disable all code which uses `std::atomic_ref` or `boost::atomic_ref` if no CPU backend is enabled
During testing the CUDA only build, I realized that the atomic_ref code was compiled all the time. The code is only used for the CPU backends. Because the check if `std::atomic_ref` is available is only enabled, if a CPU backend is enabled, the GPU backends uses the fallback `boost::atomic_ref` all the time. Therefore there was a hidden dependency on boost.